cloud_functions/pubsub/requirements.txt
Flask==2.1.1
google-cloud-pubsub==2.10.0
google-cloud-firestore==2.11.0
requests>=2.0
python-dotenv>=0.9
slack_bolt>=1.10
slack-sdk==3.21.0
redis